Servant.Server.Internal.BasicAuth

Basic Auth

data BasicAuthResult usr

servant-server's current implementation of basic authentication is not immune to certain kinds of timing attacks. Decoding payloads does not take a fixed amount of time.

The result of authentication/authorization

Constructors

Unauthorized

BadPassword

NoSuchUser

Authorized usr

newtype BasicAuthCheck usr

Datatype wrapping a function used to check authentication.

Constructors

BasicAuthCheck

Fields

unBasicAuthCheck :: BasicAuthData -> IO (BasicAuthResult usr)

mkBAChallengerHdr :: ByteString -> Header

Internal method to make a basic-auth challenge

decodeBAHdr :: Request -> Maybe BasicAuthData

Find and decode an Authorization header from the request as Basic Auth

runBasicAuth :: Request -> ByteString -> BasicAuthCheck usr -> DelayedIO usr

Run and check basic authentication, returning the appropriate http error per the spec.