Servant

Documentation

This module and its submodules can be used to define servant APIs. Note that these API definitions don't directly implement a server (or anything else).

module Servant.API

For implementing servers for servant APIs.

module Servant.Server

Utilities on top of the servant core

module Servant.Utils.Links

module Servant.Utils.StaticFiles

Useful re-exports

data Proxy k t :: forall k. k -> *

A concrete, poly-kinded proxy type

Constructors

Proxy

throwError :: MonadError e m => forall a. e -> m a

Is used within a monadic computation to begin exception processing.