Servant.Server

Description

This module lets you implement Servers for defined APIs. You'll most likely just need serve.

Run a wai application from an API

serve :: HasServer layout => Proxy layout -> Server layout -> Application

serve allows you to implement an API and produce a wai Application.

Example:

type MyApi = "books" :> Get '[JSON] [Book] -- GET /books
        :<|> "books" :> ReqBody Book :> Post '[JSON] Book -- POST /books

server :: Server MyApi
server = listAllBooks :<|> postBook
  where listAllBooks = ...
        postBook book = ...

myApi :: Proxy MyApi
myApi = Proxy

app :: Application
app = serve myApi server

main :: IO ()
main = Network.Wai.Handler.Warp.run 8080 app

Construct a wai Application from an API

toApplication :: RoutingApplication -> Application

Handlers for all standard combinators

class HasServer layout where

Associated Types

type ServerT layout m :: *

Methods

route :: Proxy layout -> Delayed (Server layout) -> Router

type Server layout = ServerT layout (ExceptT ServantErr IO)

Enter

Sometimes our cherished ExceptT monad isn't quite the type you'd like for your handlers. Maybe you want to thread some configuration in a Reader monad. Or have your types ensure that your handlers don't do any IO. Enter enter.

With enter, you can provide a function, wrapped in the `(:~>)` / Nat newtype, to convert any number of endpoints from one type constructor to another. For example

>>> import Control.Monad.Reader
>>> import qualified Control.Category as C
>>> type ReaderAPI = "ep1" :> Get '[JSON] Int :<|> "ep2" :> Get '[JSON] String
>>> let readerServer = return 1797 :<|> ask :: ServerT ReaderAPI (Reader String)
>>> let mainServer = enter (generalizeNat C.. (runReaderTNat "hi")) readerServer :: Server ReaderAPI

Basic functions and datatypes

enter :: Enter typ arg ret => arg -> typ -> ret

newtype m :~> n

A natural transformation from m to n. Used to enter particular datatypes.

Constructors

Nat

Fields

unNat :: forall a. m a -> n a

Nat utilities

liftNat :: (MonadTrans t, Monad m) => m :~> t m

Like lift.

runReaderTNat :: r -> ReaderT r m :~> m

evalStateTLNat :: Monad m => s -> StateT s m :~> m

evalStateTSNat :: Monad m => s -> StateT s m :~> m

logWriterTLNat :: MonadIO m => (w -> IO ()) -> WriterT w m :~> m

Like logWriterTSNat, but for strict WriterT.

logWriterTSNat :: MonadIO m => (w -> IO ()) -> WriterT w m :~> m

Log the contents of WriterT with the function provided as the first argument, and return the value of the WriterT computation

Functions based on mmorph

hoistNat :: (MFunctor t, Monad m) => (m :~> n) -> t m :~> t n

Like mmorph's hoist.

embedNat :: (MMonad t, Monad n) => (m :~> t n) -> t m :~> t n

Like mmorph's embed.

squashNat :: (Monad m, MMonad t) => t (t m) :~> t m

Like mmorph's squash.

generalizeNat :: Applicative m => Identity :~> m

Like mmorph's generalize.

tweakResponse :: (RouteResult Response -> RouteResult Response) -> Router -> Router

Apply a transformation to the response of a Router.

Default error type

data ServantErr

Constructors

ServantErr

Fields

errHTTPCode :: Int

errReasonPhrase :: String

errBody :: ByteString

errHeaders :: [Header]

3XX

err300 :: ServantErr

err301 :: ServantErr

err302 :: ServantErr

err303 :: ServantErr

err304 :: ServantErr

err305 :: ServantErr

err307 :: ServantErr

4XX

err400 :: ServantErr

err401 :: ServantErr

err402 :: ServantErr

err403 :: ServantErr

err404 :: ServantErr

err405 :: ServantErr

err406 :: ServantErr

err407 :: ServantErr

err409 :: ServantErr

err410 :: ServantErr

err411 :: ServantErr

err412 :: ServantErr

err413 :: ServantErr

err414 :: ServantErr

err415 :: ServantErr

err416 :: ServantErr

err417 :: ServantErr

5XX

err500 :: ServantErr

err501 :: ServantErr

err502 :: ServantErr

err503 :: ServantErr

err504 :: ServantErr

err505 :: ServantErr