Servant.Server.Internal

Documentation

type Server api = ServerT api Handler

class HasServer api context where

Associated Types

type ServerT api (m :: * -> *) :: *

The type of a server for this API, given a monad to run effects in.

Note that the result kind is *, so it is not a monad transformer, unlike what the T in the name might suggest.

Methods

route :: Proxy api -> Context context -> Delayed env (Server api) -> Router env

hoistServerWithContext :: Proxy api -> Proxy context -> (forall x. m x -> n x) -> ServerT api m -> ServerT api n

data EmptyServer

Singleton type representing a server that serves an empty API.

Constructors

EmptyServer

data AsServerT (m :: * -> *)

A type that specifies that an API record contains a server implementation.

type AsServer = AsServerT Handler

type HasServerArrowTypeError a b = ((('Text "No instance HasServer (a -> b)." ':$$: 'Text "Maybe you have used '->' instead of ':>' between ") ':$$: 'ShowType a) ':$$: 'Text "and") ':$$: 'ShowType b

type GServerConstraints api m = (ToServant api (AsServerT m) ~ ServerT (ToServantApi api) m, GServantProduct (Rep (api (AsServerT m))))

Set of constraints required to convert to / from vanilla server types.

class GServer (api :: * -> *) (m :: * -> *) where

This class is a necessary evil: in the implementation of HasServer for NamedRoutes api, we essentially need the quantified constraint forall m. GServerConstraints m to hold.

We cannot require do that directly as the definition of GServerConstraints contains type family applications (Rep and ServerT). The trick is to hide those type family applications behind a typeclass providing evidence for GServerConstraints api m in the form of a dictionary, and require that forall m. GServer api m instead.

Users shouldn't have to worry about this class, as the only possible instance is provided in this module for all record APIs.

Methods

gServerProof :: Dict (GServerConstraints api m)

emptyServer :: ServerT EmptyAPI m

Server for EmptyAPI

getAcceptHeader :: Request -> AcceptHeader

acceptCheck :: AllMime list => Proxy list -> AcceptHeader -> DelayedIO ()

allowedMethodHead :: Method -> Request -> Bool

methodCheck :: Method -> Request -> DelayedIO ()

allowedMethod :: Method -> Request -> Bool

methodRouter :: AllCTRender ctypes a => (b -> ([(HeaderName, ByteString)], a)) -> Method -> Proxy ctypes -> Status -> Delayed env (Handler b) -> Router env

noContentRouter :: Method -> Status -> Delayed env (Handler b) -> Router env

streamRouter :: forall ctype a c chunk env framing. (MimeRender ctype chunk, FramingRender framing, ToSourceIO chunk a) => (c -> ([(HeaderName, ByteString)], a)) -> Method -> Status -> Proxy framing -> Proxy ctype -> Delayed env (Handler c) -> Router env

ct_wildcard :: ByteString

module Servant.Server.Internal.BasicAuth

module Servant.Server.Internal.Context

module Servant.Server.Internal.Delayed

module Servant.Server.Internal.DelayedIO

module Servant.Server.Internal.ErrorFormatter

module Servant.Server.Internal.Handler

module Servant.Server.Internal.Router

module Servant.Server.Internal.RouteResult

module Servant.Server.Internal.RoutingApplication

module Servant.Server.Internal.ServerError